feat(gpud state): remove unused tables, clean up
